+++ docs/js/raqdps-fw-ce_ie.js
@@ -1,0 +1,19 @@
+var map = new ol.Map({
+  target: 'map',
+  layers: [
+    new ol.layer.Tile({
+      source: new ol.source.OSM()
+    }),
+    new ol.layer.Tile({
+      source: new ol.source.TileWMS({
+          format: 'image/png',
+          url: 'https://geo.weather.gc.ca/geomet/',
+          params: {'LAYERS': 'RAQDPS-FW.CE_PM2.5-DIFF-YAvg', 'TILED': true, 'STYLE': 'PM2.5_1to200ugm3'},
+      })
+    })
+  ],
+  view: new ol.View({
+    center: ol.proj.fromLonLat([-97, 57]),
+    zoom: 3
+  })
+  });